Changed Flame to Inflamable and Burns to Llamaradas.
there's a small problem with Burn word because both translations are "Quema" which is verb but also "Quema" for burn term. so if i search and replace all "Quema" it will replace both verb and noun words. just in case i mention it here if there it's a solution or smth